+++ app/Reference.php
@@ -3,6 +3,7 @@
 namespace App;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Spatie\Activitylog\Traits\LogsActivity;
 use Spatie\Activitylog\LogOptions;
 
@@ -34,7 +35,7 @@ class Reference extends Model
     ];
 
     public function getActivitylogOptions() : LogOptions
-    {
+{
         return LogOptions::defaults()
             ->logOnly(['id'])
             ->logFillable()
@@ -94,5 +95,93 @@ class Reference extends Model
 
     public function bibliography() {
         return $this->belongsTo('App\Bibliography', 'bibliography_id');
+    }
+    
+    public static function parseReferencedFromString($referenceString) {
+        $referenceString = trim($referenceString);
+        if($referenceString === '') {
+            return [];
+        }
+        $references = explode(';', $referenceString);
+        $parsedReferences = [];
+        foreach($references as $reference) {
+            $parsedReferences[] = static::parseImportReference($reference);
+        }
+        return $parsedReferences;
+    }
+    
+    public static function parseImportReference($reference) {
+        $reference = trim($reference);
+        $parts = explode(':', $reference, 2);
+        
+        $citeKey = trim($parts[0]);
+        $citeValue = (count($parts) > 1) ?  trim($parts[1]) : null;
+
+        return [
+            'valid' => count($parts) === 2,
+            'input' => $reference,
+            'citeKey' => $citeKey,
+            'description' => $citeValue,
+        ];
+    }
+    
+    public static function checkIfCiteKeyIsValid($citeKey, &$bibliographyCache) {
+            $bibliographyId = null;
+            if(isset($bibliographyCache[$citeKey])) {
+                $bibliographyId = $bibliographyCache[$citeKey];
+            } else {
+                try {
+                    $bibliography = Bibliography::where('citekey', $citeKey)->firstOrFail();
+                    $bibliographyId = $bibliography->id;
+                    $bibliographyCache[$citeKey] = $bibliographyId;
+                } catch(ModelNotFoundException $e) {
+                    $bibliographyId = null;
+                }
+            }
+            
+            return $bibliographyId != null;
+    }
+
+    public static function importReferences($referencesString, $entityId, &$bibliographyCache) {
+        $user = auth()->user();
+        $referenceString = trim($referencesString);
+        
+        if($referenceString === '') {
+            return;
+        }
+        
+        $references = explode(';', $referencesString);
+
+        foreach($references as $referenceString) {
+            
+            $reference = static::parseImportReference($referenceString);
+            $citeKey = $reference['citeKey'];
+            $citeValue = $reference['description'];
+        
+            if($reference['valid'] === false) {
+                throw new \Exception("Reference is not in the correct format. It should be in the format 'type:value' given: '$referenceString'.");
+            }
+
+            $bibliographyId = null;
+            if(isset($bibliographyCache[$citeKey])) {
+                $bibliographyId = $bibliographyCache[$citeKey];
+            } else {
+                try {
+                    $bibliography = Bibliography::where('citekey', $citeKey)->firstOrFail();
+                    $bibliographyId = $bibliography->id;
+                    $bibliographyCache[$citeKey] = $bibliographyId;
+                } catch(ModelNotFoundException $e) {
+                    throw new \Exception("Bibliography with cite key '$citeKey' does not exist.", );
+                }
+            }
+
+            Reference::firstOrCreate([
+                'entity_id' => $entityId,
+                'user_id' => $user->id,
+                'attribute_id' => null,
+                'bibliography_id' => $bibliographyId,
+                'description' => $citeValue,
+            ]);
+        }
     }
 }